requirements.txt
google-auth>=1.35.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.26.1
oauthlib>=3.1.1 
httplib2>=0.20.1
sendgrid~=6.10.0
pyjokes~=0.6.0
pytest~=8.3.4
azure-identity~=1.19.0
Markdown~=3.7
SQLAlchemy~=2.0.36 
pytest-mock~=3.14.0
pytest-socket~=0.5.1
